+++ ESI.NET/Models/PlanetaryInteraction/CustomsOffice.cs
@@ -18,13 +18,13 @@ namespace ESI.NET.Models.PlanetaryInteraction
         public decimal BadStandingTaxRate { get; set; }
 
         [JsonProperty("corporation_tax_rate")]
-        public string CorporationTaxRate { get; set; }
+        public decimal CorporationTaxRate { get; set; }
 
         [JsonProperty("excellent_standing_tax_rate")]
-        public long ExcellentStandingTaxRate { get; set; }
+        public decimal ExcellentStandingTaxRate { get; set; }
 
         [JsonProperty("good_standing_tax_rate")]
-        public int GoodStandingTaxRate { get; set; }
+        public decimal GoodStandingTaxRate { get; set; }
 
         [JsonProperty("neutral_standing_tax_rate")]
         public decimal NeutralStandingTaxRate { get; set; }
@@ -46,5 +46,8 @@ namespace ESI.NET.Models.PlanetaryInteraction
 
         [JsonProperty("terrible_standing_tax_rate")]
         public decimal TerribleDtandingRate { get; set; }
+
+        [JsonProperty("type_id")]
+        public long TypeId { get; set; }
     }
 }